Rename to batch-open-issues; add migration
