+++ source/background.js
@@ -16,12 +16,12 @@ new OptionsSync().define({
 			options.disabledFeatures = options.disabledFeatures.replace('toggle-all-things-with-alt', ''); // #1524
 			options.disabledFeatures = options.disabledFeatures.replace('remove-diff-signs', ''); // #1524
 			options.disabledFeatures = options.disabledFeatures.replace('add-confirmation-to-comment-cancellation', ''); // #1524
-
 			options.disabledFeatures = options.disabledFeatures.replace('add-your-repositories-link-to-profile-dropdown', ''); // #1460
 			options.disabledFeatures = options.disabledFeatures.replace('add-readme-buttons', 'hide-readme-header'); // #1465
 			options.disabledFeatures = options.disabledFeatures.replace('add-delete-to-pr-files', ''); // #1462
 			options.disabledFeatures = options.disabledFeatures.replace('auto-load-more-news', 'infinite-scroll'); // #1516
 			options.disabledFeatures = options.disabledFeatures.replace('display-issue-suggestions', ''); // #1611
+			options.disabledFeatures = options.disabledFeatures.replace('open-all-selected', 'batch-open-issues'); // #1402
 		},
 		OptionsSync.migrations.removeUnused
 	]
